fix(misc): remove leftover code from initial approach and remove unnecessary whitespace from jest config
