+++ packages/jest/src/generators/configuration/lib/create-jest-config.ts
@@ -32,7 +32,7 @@ export default { ...nxPreset };`
       tree.write(
         `jest.preset.${presetExt}`,
         `const nxPreset = require('@nx/jest/preset').default;
-  
+
 module.exports = { ...nxPreset };`
       );
     }
